Trading View Chart

+++ TradeForge.Core/Models/OHLC.cs
@@ -3,9 +3,9 @@
 public class OHLC
 {
     public DateTime Timestamp { get; set; } = DateTime.MinValue;
-    public decimal Open { get; set; } = 0;
-    public decimal High { get; set; } = 0;
-    public decimal Low { get; set; } = 0;
-    public decimal Close { get; set; } = 0;
-    public decimal Volume { get; set; } = 0;
+    public double Open { get; set; } = 0;
+    public double High { get; set; } = 0;
+    public double Low { get; set; } = 0;
+    public double Close { get; set; } = 0;
+    public double Volume { get; set; } = 0;
 }
